+++ jquery.barrating.js
@@ -1,5 +1,5 @@
 /**
- * jQuery Bar Rating Plugin v1.0.6
+ * jQuery Bar Rating Plugin v1.1
  *
  * http://github.com/antennaio/jquery-bar-rating
  *