+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-HEAD
-----
+1.1
+---
 
 - new & improved demo page